+++ front/app/modules/commercial/content_builder/admin/components/ContentBuilderPreview/index.tsx
@@ -1,27 +1,21 @@
 import React, { useEffect } from 'react';
 import { withRouter, WithRouterProps } from 'react-router';
 
-// Hooks
+// hooks
 import useProject from 'hooks/useProject';
 import useFeatureFlag from 'hooks/useFeatureFlag';
-import useContentBuilderLayout from '../../../hooks/useContentBuilder';
-import useLocale from 'hooks/useLocale';
 
-// Components
-import Editor from '../Editor';
-import ContentBuilderFrame from '../ContentBuilderFrame';
-import { Box } from '@citizenlab/cl2-component-library';
-import ProjectInfo from 'containers/ProjectsShowPage/shared/header/ProjectInfo';
+// components
+import ContentPreview from './ContentPreview';
+
+// utils
 import { isNilOrError } from 'utils/helperUtils';
-
-// services
-import { PROJECT_DESCRIPTION_CODE } from '../../../services/contentBuilder';
 
 type ContentBuilderPreviewProps = {
   onMount: () => void;
 } & WithRouterProps;
 
-const ContentBuilderPreviewContainer = ({
+const ContentBuilderPreview = ({
   onMount,
   params: { slug },
 }: ContentBuilderPreviewProps) => {
@@ -39,34 +33,14 @@ const ContentBuilderPreviewContainer = ({
 
   if (isNilOrError(project)) return null;
 
-  return <ContentBuilderPreview projectId={project.id} />;
-};
-
-const ContentBuilderPreview = ({ projectId }: { projectId: string }) => {
-  const locale = useLocale();
-
-  const contentBuilderLayout = useContentBuilderLayout({
-    projectId,
-    code: PROJECT_DESCRIPTION_CODE,
-  });
-
-  if (isNilOrError(locale)) return null;
-
-  const contentBuilderContent =
-    !isNilOrError(contentBuilderLayout) &&
-    contentBuilderLayout.data.attributes.craftjs_jsonmultiloc[locale];
-
-  console.log(contentBuilderContent);
   return (
-    <Box>
-      {contentBuilderContent ? (
-        <Editor isPreview={true}>
-          <ContentBuilderFrame projectId={projectId} />
-        </Editor>
-      ) : (
-        <ProjectInfo projectId={projectId} />
-      )}
-    </Box>
+    <>
+      <ContentPreview
+        projectId={project.id}
+        projectTitle={project.attributes.title_multiloc}
+      />
+    </>
   );
 };
-export default withRouter(ContentBuilderPreviewContainer);
+
+export default withRouter(ContentBuilderPreview);
